Added support for distinguishing file versions.
When splitting, the target version is determined by the source file.
Data fields that are not relevant for this version may be left out.

When combining, the target version is determined by the destination
file extension. All data fields that are present will be read,
but warnings will be issued if the target format is not able to
represent them.