Fix #9: References to arguments of Drag'n'Drop actions could be lost when splitting
This is because LateralGM created a list of postponed resolution actions if it could not immediately resolve a reference while reading a .gmk or .gm81 file, but then forgot to do anything with that list afterwards.

This was fixed shortly later in https://github.com/IsmAvatar/LateralGM/commit/c1826a829f1ebc9751015d05c9c15f87aa1488b9 so I could have updated to a more recent verison of LateralGM instead, but since I feel that could be fairly involved and don't have much time right now this workaround will have to do.